+++ orttraining/orttraining/core/graph/optimizer/sgd_optimizer_builder.h
@@ -46,7 +46,7 @@ class SGDOptimizerBuilder final : public OptimizerBuilder {
       std::vector<ArgDef>& output_weight_argdefs,
       std::vector<ArgDef>& output_gradient_argdefs,
       const bool enable_grad_clipping,
-      const NameMLValMap shared_optim_state) const override;
+      const NameMLValMap& shared_optim_state) const override;
 };
 
 }  // namespace training
