+++ orttraining/orttraining/test/graph/gradient_graph_builder_test.cc
@@ -11,7 +11,7 @@
 #include "core/session/environment.h"
 #include "orttraining/core/framework/distributed_run_context.h"
 #include "orttraining/models/runner/training_runner.h"
-#include "orttraining/test/graph/training_session_test_utils.h"
+#include "orttraining/test/session/training_session_test_utils.h"
 
 #include "orttraining/training_ops/cpu/controlflow/event_pool.h"  // TODO: move with PipelineBatchPlanner
 
@@ -24,9 +24,92 @@ using namespace onnxruntime::logging;
 using namespace onnxruntime::training;
 using namespace google::protobuf::util;
 using namespace onnxruntime::path_utils;
+using namespace onnxruntime::test::training_session_test_utils;
 
 namespace onnxruntime {
 namespace test {
+
+namespace {
+constexpr auto CONCAT_MODEL_PATH = ORT_TSTR("testdata/transform/concat_trainable.onnx");
+}  // namespace
+
+static Status BuildBackPropGraph(
+    const PathString& forward_model_file,
+    const TrainingSession::TrainingConfiguration& config,
+    PathString& backward_model_file) {
+  std::unique_ptr<Environment> env;
+  ORT_RETURN_IF_ERROR(Environment::Create(nullptr, env));
+
+  SessionOptions so{};
+  TrainingSession training_session{so, *env};
+
+  std::cout << "Loading source model file = " << ToMBString(forward_model_file) << "\n";
+
+  ORT_RETURN_IF_ERROR(training_session.Load(forward_model_file));
+
+  TrainingSession::TrainingConfigurationResult config_result{};
+  ORT_RETURN_IF_ERROR(training_session.ConfigureForTraining(config, config_result));
+
+  backward_model_file = config.model_with_training_graph_path.value();
+
+  return Status::OK();
+}
+
+/**
+ * Run a training session for this model for 1 epoch, using batch size of 1 and synthetic input data.
+ * @param so - SessionOptions for this run.
+ * @param backprop_model_file - Model file to be run. This should already contain loss function and backward prop nodes.
+ * @return TrainingSession for this run.
+ */
+static std::unique_ptr<TrainingSession> RunTrainingSessionWithChecks(
+    const SessionOptions& so, const PathString& backprop_model_file) {
+  std::unique_ptr<Environment> env;
+  ORT_THROW_IF_ERROR(Environment::Create(nullptr, env));
+
+  std::unique_ptr<TrainingSession> training_session = onnxruntime::make_unique<TrainingSession>(so, *env);
+
+  ORT_THROW_IF_ERROR(training_session->Load(backprop_model_file));
+
+  std::pair<common::Status, const ModelMetadata*> res = training_session->GetModelMetadata();
+  ORT_THROW_IF_ERROR(res.first);
+  ORT_ENFORCE(res.second != nullptr);
+  auto model_metadata = res.second;
+  std::cout << "Loaded " << model_metadata->graph_name << '\n';
+
+  ORT_THROW_IF_ERROR(training_session->Initialize());
+
+  std::vector<MLValue> gradient_fetches;
+  RunOptions run_options;
+  run_options.run_log_verbosity_level = so.session_log_verbosity_level;
+  run_options.run_tag = so.session_logid;
+  run_options.training_mode = true;
+
+  // Create dummy feeds
+  std::vector<int64_t> image_dims = {1, 784};
+  std::vector<int64_t> label_dims = {1, 10};
+  std::vector<float> image_value(784, 1);
+  std::vector<float> label_value(10, 1);
+
+  MLValue imageMLValue;
+  TrainingUtil::CreateCpuMLValue(image_dims, image_value, &imageMLValue);
+  MLValue labelMLValue;
+  TrainingUtil::CreateCpuMLValue(label_dims, label_value, &labelMLValue);
+
+  auto fw_feeds = std::make_pair<std::vector<std::string>, std::vector<MLValue>>({"X", "labels"}, {imageMLValue, labelMLValue});
+
+  auto output_names_include_gradients = GetModelOutputNames(*training_session);
+  std::vector<std::string> training_output_names(output_names_include_gradients.begin(), output_names_include_gradients.end());
+
+  auto start_time = std::chrono::high_resolution_clock::now();
+
+  ORT_THROW_IF_ERROR(training_session->Run(run_options, fw_feeds.first, fw_feeds.second, training_output_names, &gradient_fetches));
+
+  auto end_time = std::chrono::high_resolution_clock::now();
+  auto elapsed = TimeDiffMicroSeconds(start_time, end_time);
+  std::cout << "Training session run completed in " << elapsed << " microseconds.\n";
+
+  return training_session;
+}
 
 TEST(GradientGraphBuilderTest, BuildGradientGraphTest) {
   const auto config = MakeBasicTrainingConfig();
@@ -216,6 +299,172 @@ TEST(GradientGraphBuilderTest, TrainingSession_WithProfiler) {
   ASSERT_TRUE(count > 1);
 }
 
+#ifdef USE_CUDA
+static void RunBertTrainingWithChecks(
+    const SessionOptions& so,
+    const PathString& backprop_model_file) {
+  std::unique_ptr<Environment> env;
+  ASSERT_STATUS_OK(Environment::Create(nullptr, env));
+
+  std::unique_ptr<TrainingSession> training_session = onnxruntime::make_unique<TrainingSession>(so, *env);
+
+  ASSERT_STATUS_OK(training_session->Load(backprop_model_file));
+
+  std::pair<common::Status, const ModelMetadata*> res = training_session->GetModelMetadata();
+  ASSERT_STATUS_OK(res.first);
+  ASSERT_TRUE(res.second != nullptr);
+  auto model_metadata = res.second;
+  std::cout << "Loaded " << model_metadata->graph_name << '\n';
+
+  CUDAExecutionProviderInfo xp_info;
+  ASSERT_STATUS_OK(training_session->RegisterExecutionProvider(onnxruntime::make_unique<CUDAExecutionProvider>(xp_info)));
+
+  ASSERT_STATUS_OK(training_session->Initialize());
+
+  RunOptions run_options;
+  run_options.run_log_verbosity_level = so.session_log_verbosity_level;
+  run_options.run_tag = so.session_logid;
+  run_options.training_mode = true;
+
+  // Creating feeds
+  int batch_size = 13;
+  int max_seq_len_in_batch = 7;
+  std::vector<std::string> feed_names = {
+      "input_ids",
+      "token_type_ids",
+      "input_mask",
+      "masked_lm_ids",
+      "next_sentence_labels",
+      "masked_lm_positions",
+  };
+  std::vector<TensorShape> tensor_shapes = {
+      {batch_size, max_seq_len_in_batch},
+      {batch_size, max_seq_len_in_batch},
+      {batch_size, max_seq_len_in_batch},
+      {batch_size, max_seq_len_in_batch},
+      {batch_size},
+      {batch_size, max_seq_len_in_batch},
+      {batch_size, max_seq_len_in_batch}};
+
+  std::vector<std::vector<int64_t>> tensor_values = {
+      /*input_ids*/
+      {49, 97, 53, 5, 33, 65, 62,
+       51, 38, 61, 45, 74, 27, 64,
+       17, 36, 17, 96, 12, 79, 32,
+       68, 90, 77, 18, 39, 12, 93,
+       9, 87, 42, 60, 71, 12, 45,
+       55, 40, 78, 81, 26, 70, 61,
+       56, 66, 33, 7, 70, 1, 11,
+       92, 51, 90, 85, 80, 0, 78,
+       63, 42, 31, 93, 41, 90, 8,
+       24, 72, 28, 30, 18, 69, 57,
+       11, 10, 40, 65, 62, 13, 38,
+       70, 37, 90, 15, 70, 42, 69,
+       26, 77, 70, 75, 36, 56, 11},
+      /*token_type_ids*/
+      {12, 13, 1, 8, 15, 12, 9,
+       15, 11, 6, 4, 9, 4, 3,
+       8, 4, 9, 3, 2, 10, 15,
+       3, 11, 13, 10, 6, 15, 14,
+       8, 1, 0, 2, 12, 0, 15,
+       10, 7, 10, 2, 6, 7, 7,
+       4, 14, 2, 2, 10, 15, 3,
+       9, 9, 3, 10, 6, 9, 14,
+       2, 12, 10, 7, 9, 5, 6,
+       5, 1, 8, 15, 2, 2, 4,
+       4, 1, 2, 12, 8, 7, 6,
+       13, 8, 14, 15, 11, 2, 10,
+       3, 15, 10, 6, 7, 0, 8},
+      /*input_mask*/
+      {1, 1, 0, 1, 1, 1, 1,
+       1, 1, 0, 0, 1, 0, 0,
+       1, 0, 1, 0, 0, 1, 1,
+       0, 1, 1, 1, 0, 1, 1,
+       1, 0, 0, 0, 1, 0, 1,
+       1, 0, 1, 0, 0, 0, 0,
+       0, 1, 0, 0, 1, 1, 0,
+       1, 1, 0, 1, 0, 1, 1,
+       0, 1, 1, 0, 1, 0, 0,
+       0, 0, 1, 1, 0, 0, 0,
+       0, 0, 0, 1, 1, 0, 0,
+       1, 1, 1, 1, 1, 0, 1,
+       0, 1, 1, 0, 0, 0, 1},
+      /*masked_lm_ids*/
+      {1, 1, 0, 1, 2, 1, 1,
+       1, 1, 1, 2, 0, 2, 0,
+       1, 0, 0, 2, 1, 2, 2,
+       2, 0, 1, 0, 2, 0, 2,
+       1, 1, 2, 0, 1, 1, 1,
+       2, 2, 0, 2, 1, 1, 2,
+       1, 0, 2, 0, 0, 2, 1,
+       2, 2, 2, 0, 2, 1, 1,
+       0, 2, 1, 2, 0, 0, 2,
+       0, 0, 0, 2, 1, 0, 0,
+       1, 2, 1, 0, 1, 2, 1,
+       2, 0, 2, 1, 2, 0, 2,
+       2, 2, 1, 1, 0, 2, 1},
+      /*next_sentence_labels*/
+      {1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 0},
+      /*masked_lm_positions*/
+      {0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6,
+       0, 1, 2, 3, 4, 5, 6}};
+
+  std::vector<OrtValue> feeds(feed_names.size());
+  for (size_t i = 0; i < 6; ++i) {
+    TrainingUtil::CreateCpuMLValue(tensor_shapes[i].GetDims(), tensor_values[i], &feeds[i]);
+  }
+
+  auto output_names_include_gradients = GetModelOutputNames(*training_session);
+  std::vector<std::string> fetch_names(output_names_include_gradients.begin(), output_names_include_gradients.end());
+
+  std::vector<OrtValue> fetches;
+
+  ASSERT_STATUS_OK(training_session->Run(run_options, feed_names, feeds, fetch_names, &fetches));
+
+  for (size_t i = 0; i < fetch_names.size(); ++i) {
+    if (!fetches[i].IsAllocated() || !!fetches[i].IsTensor())
+      continue;
+
+    const Tensor& tensor = fetches[i].Get<Tensor>();
+    if (DataTypeImpl::GetType<float>() != tensor.DataType()) {
+      continue;
+    }
+
+    const std::string& name = fetch_names[i];
+    if (BERT_TOY_FETCHES.find(name) == BERT_TOY_FETCHES.end()) {
+      continue;
+    }
+
+    auto gradient_ref = BERT_TOY_FETCHES.at(name);
+    EXPECT_TRUE(static_cast<size_t>(tensor.Shape().Size()) == gradient_ref.size());
+
+    float max_diff = 0;
+    float max_percent_diff = 0;
+    const float* data = tensor.template Data<float>();
+    for (size_t idx = 0; idx < gradient_ref.size(); ++idx) {
+      float diff = std::fabs(static_cast<float>(gradient_ref[idx]) - data[idx]);
+      max_diff = std::fmax(max_diff, diff);
+      max_percent_diff = std::fmax(max_percent_diff, diff / data[idx]);
+    }
+    EXPECT_TRUE(max_diff < 1e-5) << name << " is incorrect: max_diff is " << max_diff;
+    if (max_diff > 1e-10) {
+      EXPECT_TRUE(max_percent_diff < 0.01f) << name << " is incorrect: max_percent_diff is "
+                                            << max_percent_diff;
+    }
+  }
+}
+#endif
 TEST(GradientGraphBuilderTest, TrainingSession_BertToy) {
   const auto model_path = ORT_TSTR("testdata/bert_toy_optimized.onnx");
